Refactor: Deleted unnecessary nested folder
